+++ lib/screens/numchkpage/widget/animation_container.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 import 'package:pcg_pos/widget/MyElevatedBtn.dart';
-import 'package:pcg_pos/widget/app_bar.dart';
 
 class NumCheck extends StatefulWidget {
   final String title;
@@ -17,84 +16,71 @@ class NumCheck extends StatefulWidget {
 
 class _NumCheckState extends State<NumCheck> {
   bool _isExpanded = false;
-  bool _isEntry = true;
-
-  void _toggleEntryExit() {
-    setState(() {
-      _isEntry = !_isEntry;
-    });
-  }
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: MyAppBar(
-          title: widget.title,
-          isEntry: _isEntry,
-          onToggle: _toggleEntryExit,
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.center,
+      children: [
+        SizedBox(
+          height: MediaQuery.of(context).size.height * 0.15,
         ),
-        body: Column(
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            SizedBox(
-              height: MediaQuery.of(context).size.height * 0.15,
-            ),
-            AnimatedContainer(
-              duration: const Duration(milliseconds: 300),
-              width: MediaQuery.of(context).size.width * 1,
-              height: _isExpanded
-                  ? MediaQuery.of(context).size.height * 0.2
-                  : MediaQuery.of(context).size.height * 0.4,
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(10),
-              ),
-              child: _isExpanded
-                  ? Container(
-                      width: MediaQuery.of(context).size.height * 0.2,
-                      padding: const EdgeInsets.symmetric(horizontal: 20),
-                      decoration: BoxDecoration(
-                          border: Border.all(
-                            color: const Color(0xff39c5bb),
-                            width: 4,
-                          ),
-                          borderRadius: BorderRadius.circular(20)),
-                      child: const TextField(
-                        maxLength: 8,
-                        maxLengthEnforcement: MaxLengthEnforcement.enforced,
-                        decoration: InputDecoration(
-                          counterText: '',
-                          border: InputBorder.none,
-                        ),
-                        style: TextStyle(
-                          letterSpacing: 20,
-                          fontSize: 60,
-                        ),
+        AnimatedContainer(
+          duration: const Duration(milliseconds: 300),
+          width: MediaQuery.of(context).size.width * 1,
+          height: _isExpanded
+              ? MediaQuery.of(context).size.height * 0.2
+              : MediaQuery.of(context).size.height * 0.4,
+          decoration: BoxDecoration(
+            borderRadius: BorderRadius.circular(10),
+          ),
+          child: _isExpanded
+              ? Container(
+                  width: MediaQuery.of(context).size.height * 0.2,
+                  padding: const EdgeInsets.symmetric(horizontal: 20),
+                  decoration: BoxDecoration(
+                      border: Border.all(
+                        color: const Color(0xff39c5bb),
+                        width: 4,
                       ),
-                    )
-                  : const CarNumber(),
-            ),
-            SizedBox(
-              height: MediaQuery.of(context).size.height * 0.1,
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                MyElevatedBtn(
-                  text: "아니오",
-                  onPressed: () {
-                    Navigator.pushNamed(context, '/');
-                  },
-                ),
-                MyElevatedBtn(
-                  text: "예",
-                  onPressed: () {
-                    Navigator.pushNamed(context, '/');
-                  },
+                      borderRadius: BorderRadius.circular(20)),
+                  child: const TextField(
+                    maxLength: 8,
+                    maxLengthEnforcement: MaxLengthEnforcement.enforced,
+                    decoration: InputDecoration(
+                      counterText: '',
+                      border: InputBorder.none,
+                    ),
+                    style: TextStyle(
+                      letterSpacing: 20,
+                      fontSize: 60,
+                    ),
+                  ),
                 )
-              ],
+              : const CarNumber(),
+        ),
+        SizedBox(
+          height: MediaQuery.of(context).size.height * 0.1,
+        ),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            MyElevatedBtn(
+              text: "아니오",
+              onPressed: () {
+                Navigator.pushReplacementNamed(context, '/');
+              },
+            ),
+            MyElevatedBtn(
+              text: "예",
+              onPressed: () {
+                Navigator.pushReplacementNamed(context, '/');
+              },
             )
           ],
-        ));
+        )
+      ],
+    );
   }
 
   void expandControl() {
@@ -130,12 +116,10 @@ class CarNumber extends StatelessWidget {
       ),
       height: MediaQuery.of(context).size.height * 0.40,
       child: const Center(
-        child: TextField(
-            style: TextStyle(fontSize: 50),
-            decoration: InputDecoration(
-              border: OutlineInputBorder(),
-              labelText: '차량 번호 입력',
-            )),
+        child: Text(
+          "       107소 2586\n이 번호가 맞습니까?", //TODO: 하드코딩
+          style: TextStyle(fontSize: 50),
+        ),
       ),
     );
   }
